Restyle to match airborn
Since we can't use javascript on the homepage, we use a static
background there instead of the rotating one. Generated with product=58.

Removes the "cut-out" effect in docs/ on mobile and tablets. It now
requires background-attachment: fixed. Various webkits don't support
that, and Firefox Mobile's implementation is not smooth.

Also fixes small issues with the logo on docs/.